+++ alacritty_terminal/Cargo.toml
@@ -23,7 +23,6 @@ serde_yaml = "0.8"
 vte = "0.3"
 mio = "0.6"
 mio-extras = "2"
-xdg = "2"
 log = "0.4"
 fnv = "1"
 unicode-width = "0.1"
@@ -48,7 +47,6 @@ mio-named-pipes = "0.1"
 miow = "0.3"
 dunce = "1.0"
 winapi = { version = "0.3.7", features = ["impl-default", "winuser", "synchapi", "roerrorapi", "winerror", "wincon", "wincontypes"]}
-dirs = "1.0"
 widestring = "0.4"
 mio-anonymous-pipes = "0.1"
