+++ alacritty_terminal/src/term/color.rs
@@ -1,5 +1,9 @@
 use std::fmt;
 use std::ops::{Index, IndexMut, Mul};
+use std::str::FromStr;
+
+use serde::de::Visitor;
+use serde::{Deserialize, Deserializer};
 
 use crate::ansi;
 use crate::config::Colors;
@@ -9,7 +13,7 @@ pub const COUNT: usize = 270;
 pub const RED: Rgb = Rgb { r: 0xff, g: 0x0, b: 0x0 };
 pub const YELLOW: Rgb = Rgb { r: 0xff, g: 0xff, b: 0x0 };
 
-#[derive(Debug, Eq, PartialEq, Copy, Clone, Default, Serialize, Deserialize)]
+#[derive(Debug, Eq, PartialEq, Copy, Clone, Default, Serialize)]
 pub struct Rgb {
     pub r: u8,
     pub g: u8,
@@ -30,6 +34,99 @@ impl Mul<f32> for Rgb {
         trace!("Scaling RGB by {} from {:?} to {:?}", rhs, self, result);
 
         result
+    }
+}
+
+/// Deserialize an Rgb from a hex string
+///
+/// This is *not* the deserialize impl for Rgb since we want a symmetric
+/// serialize/deserialize impl for ref tests.
+impl<'de> Deserialize<'de> for Rgb {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        struct RgbVisitor;
+
+        // Used for deserializing reftests
+        #[derive(Deserialize)]
+        struct RgbDerivedDeser {
+            r: u8,
+            g: u8,
+            b: u8,
+        }
+
+        impl<'a> Visitor<'a> for RgbVisitor {
+            type Value = Rgb;
+
+            fn expecting(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+                f.write_str("hex color like 0xff00ff")
+            }
+
+            fn visit_str<E>(self, value: &str) -> ::std::result::Result<Rgb, E>
+            where
+                E: ::serde::de::Error,
+            {
+                Rgb::from_str(&value[..])
+                    .map_err(|_| E::custom("failed to parse rgb; expected hex color like 0xff00ff"))
+            }
+        }
+
+        // Return an error if the syntax is incorrect
+        let value = serde_yaml::Value::deserialize(deserializer)?;
+
+        // Attempt to deserialize from struct form
+        if let Ok(RgbDerivedDeser { r, g, b }) = RgbDerivedDeser::deserialize(value.clone()) {
+            return Ok(Rgb { r, g, b });
+        }
+
+        // Deserialize from hex notation (either 0xff00ff or #ff00ff)
+        match value.deserialize_str(RgbVisitor) {
+            Ok(rgb) => Ok(rgb),
+            Err(err) => {
+                error!("Problem with config: {}; using color #000000", err);
+                Ok(Rgb::default())
+            },
+        }
+    }
+}
+
+impl FromStr for Rgb {
+    type Err = ();
+
+    fn from_str(s: &str) -> ::std::result::Result<Rgb, ()> {
+        let mut chars = s.chars();
+        let mut rgb = Rgb::default();
+
+        macro_rules! component {
+            ($($c:ident),*) => {
+                $(
+                    match chars.next().and_then(|c| c.to_digit(16)) {
+                        Some(val) => rgb.$c = (val as u8) << 4,
+                        None => return Err(())
+                    }
+
+                    match chars.next().and_then(|c| c.to_digit(16)) {
+                        Some(val) => rgb.$c |= val as u8,
+                        None => return Err(())
+                    }
+                )*
+            }
+        }
+
+        match chars.next() {
+            Some('0') => {
+                if chars.next() != Some('x') {
+                    return Err(());
+                }
+            },
+            Some('#') => (),
+            _ => return Err(()),
+        }
+
+        component!(r, g, b);
+
+        Ok(rgb)
     }
 }
 
@@ -60,24 +157,24 @@ impl<'a> From<&'a Colors> for List {
 impl List {
     pub fn fill_named(&mut self, colors: &Colors) {
         // Normals
-        self[ansi::NamedColor::Black] = colors.normal.black;
-        self[ansi::NamedColor::Red] = colors.normal.red;
-        self[ansi::NamedColor::Green] = colors.normal.green;
-        self[ansi::NamedColor::Yellow] = colors.normal.yellow;
-        self[ansi::NamedColor::Blue] = colors.normal.blue;
-        self[ansi::NamedColor::Magenta] = colors.normal.magenta;
-        self[ansi::NamedColor::Cyan] = colors.normal.cyan;
-        self[ansi::NamedColor::White] = colors.normal.white;
+        self[ansi::NamedColor::Black] = colors.normal().black;
+        self[ansi::NamedColor::Red] = colors.normal().red;
+        self[ansi::NamedColor::Green] = colors.normal().green;
+        self[ansi::NamedColor::Yellow] = colors.normal().yellow;
+        self[ansi::NamedColor::Blue] = colors.normal().blue;
+        self[ansi::NamedColor::Magenta] = colors.normal().magenta;
+        self[ansi::NamedColor::Cyan] = colors.normal().cyan;
+        self[ansi::NamedColor::White] = colors.normal().white;
 
         // Brights
-        self[ansi::NamedColor::BrightBlack] = colors.bright.black;
-        self[ansi::NamedColor::BrightRed] = colors.bright.red;
-        self[ansi::NamedColor::BrightGreen] = colors.bright.green;
-        self[ansi::NamedColor::BrightYellow] = colors.bright.yellow;
-        self[ansi::NamedColor::BrightBlue] = colors.bright.blue;
-        self[ansi::NamedColor::BrightMagenta] = colors.bright.magenta;
-        self[ansi::NamedColor::BrightCyan] = colors.bright.cyan;
-        self[ansi::NamedColor::BrightWhite] = colors.bright.white;
+        self[ansi::NamedColor::BrightBlack] = colors.bright().black;
+        self[ansi::NamedColor::BrightRed] = colors.bright().red;
+        self[ansi::NamedColor::BrightGreen] = colors.bright().green;
+        self[ansi::NamedColor::BrightYellow] = colors.bright().yellow;
+        self[ansi::NamedColor::BrightBlue] = colors.bright().blue;
+        self[ansi::NamedColor::BrightMagenta] = colors.bright().magenta;
+        self[ansi::NamedColor::BrightCyan] = colors.bright().cyan;
+        self[ansi::NamedColor::BrightWhite] = colors.bright().white;
         self[ansi::NamedColor::BrightForeground] =
             colors.primary.bright_foreground.unwrap_or(colors.primary.foreground);
 
@@ -106,14 +203,14 @@ impl List {
             },
             None => {
                 trace!("Deriving dim colors from normal colors");
-                self[ansi::NamedColor::DimBlack] = colors.normal.black * 0.66;
-                self[ansi::NamedColor::DimRed] = colors.normal.red * 0.66;
-                self[ansi::NamedColor::DimGreen] = colors.normal.green * 0.66;
-                self[ansi::NamedColor::DimYellow] = colors.normal.yellow * 0.66;
-                self[ansi::NamedColor::DimBlue] = colors.normal.blue * 0.66;
-                self[ansi::NamedColor::DimMagenta] = colors.normal.magenta * 0.66;
-                self[ansi::NamedColor::DimCyan] = colors.normal.cyan * 0.66;
-                self[ansi::NamedColor::DimWhite] = colors.normal.white * 0.66;
+                self[ansi::NamedColor::DimBlack] = colors.normal().black * 0.66;
+                self[ansi::NamedColor::DimRed] = colors.normal().red * 0.66;
+                self[ansi::NamedColor::DimGreen] = colors.normal().green * 0.66;
+                self[ansi::NamedColor::DimYellow] = colors.normal().yellow * 0.66;
+                self[ansi::NamedColor::DimBlue] = colors.normal().blue * 0.66;
+                self[ansi::NamedColor::DimMagenta] = colors.normal().magenta * 0.66;
+                self[ansi::NamedColor::DimCyan] = colors.normal().cyan * 0.66;
+                self[ansi::NamedColor::DimWhite] = colors.normal().white * 0.66;
             },
         }
     }
